update 2
big changes, alot of javascript added to make product flow somewhat seemless.
fixed and/or improve alot of css to match this new data. may have turned alot of line of code redundant, but havent had the time to go through and fix that.
this commit is to get the page on the web to fine tune testing so that it works not only localy.